XWIKI-22683: Move org.xwiki.security.authorization.DefaultAuthorizationManager to internal package
  * Move DefaultAuthorizationManager to internal package
  * Expose some APIs that were package protected to make them public so
    that the component implementation can access them
  * Add a revapi ignore to explain the removed public API
