+++ xwiki-platform-core/xwiki-platform-security/xwiki-platform-security-authorization/xwiki-platform-security-authorization-bridge/src/main/java/org/xwiki/security/authorization/internal/BridgeAuthorizationManager.java
@@ -27,7 +27,7 @@ import org.xwiki.model.reference.DocumentReference;
 import org.xwiki.model.reference.EntityReference;
 import org.xwiki.rendering.async.AsyncContext;
 import org.xwiki.security.authorization.AccessDeniedException;
-import org.xwiki.security.authorization.DefaultAuthorizationManager;
+import org.xwiki.security.internal.authorization.DefaultAuthorizationManager;
 import org.xwiki.security.authorization.Right;
 
 /**
